refactor: rename app/packages for consistancy
