+++ dev/com.ibm.ws.concurrent_fat_cdi/.classpath
@@ -5,7 +5,7 @@
 	<classpathentry kind="src" path="test-applications/concurrentCDI3App/src"/>
 	<classpathentry kind="src" path="test-applications/concurrentCDI4App/src"/>
 	<classpathentry kind="src" path="test-applications/concurrentCDIWeb/src"/>
-	<classpathentry kind="src" path="test-applications/concurrentCDIWeb2/src"/>
+	<classpathentry kind="src" path="test-applications/concurrentCDI4WebApp/src"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
 		<attributes>
