+++ dev/com.ibm.ws.concurrent_fat_cdi/bnd.bnd
@@ -18,8 +18,8 @@ src:\
     test-applications/concurrentCDI2App/src,\
     test-applications/concurrentCDI3App/src,\
     test-applications/concurrentCDI4App/src,\
-    test-applications/concurrentCDIWeb/src,\
-    test-applications/concurrentCDIWeb2/src
+    test-applications/concurrentCDI4WebApp/src,\
+    test-applications/concurrentCDIWeb/src
 
 fat.project: true
