+++ dev/com.ibm.ws.concurrent_fat_cdi/test-applications/concurrentCDI4WebApp/resources/WEB-INF/beans.xml
