+++ dev/com.ibm.ws.concurrent_fat_cdi/test-applications/concurrentCDI4WebApp/src/concurrent/cdi4/webapp/TestBean.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package concurrent.cdi.web2;
+package concurrent.cdi4.webapp;
 
 import static org.junit.Assert.assertNotNull;
